New GUI controls added.   - Added the ability to (re)set a background color within the GUI;   - Added the ability to (re)set the pressure factor within the GUI;

+++ src/datatypes/configuration.h
@@ -24,7 +24,7 @@ typedef struct
 {
   unsigned char type;
   int num_colors;
-  float pressure_factor;
+  double pressure_factor;
   char** colors;
   char* background;
 } dt_configuration;
